# backend/src/main/java/com/turkcell/ecommerce/category/dto/CategoryResponse.java
package com.turkcell.ecommerce.category.dto;

import lombok.*;

import java.time.LocalDateTime;
import java.util.UUID;

@Getter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CategoryResponse {

    private UUID id;
    private String name;
    private String description;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
}