# docs/openapi/category-v1.yml
openapi: 3.0.3
info:
  title: Kategori Servisi
  description: Mini E-Ticaret uygulaması için Kategori CRUD API'si
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1

tags:
  - name: Categories
    description: Kategori yönetim işlemleri

paths:
  /categories:
    get:
      tags:
        - Categories
      summary: Tüm kategorileri listele
      operationId: getAllCategories
      responses:
        "200":
          description: Kategoriler başarıyla listelendi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Categories
      summary: Yeni kategori oluştur
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "201":
          description: Kategori başarıyla oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Kategori UUID
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Categories
      summary: ID ile kategori getir
      operationId: getCategoryById
      responses:
        "200":
          description: Kategori başarıyla getirildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags:
        - Categories
      summary: Kategoriyi güncelle
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "200":
          description: Kategori başarıyla güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Categories
      summary: Kategoriyi sil
      operationId: deleteCategory
      responses:
        "204":
          description: Kategori başarıyla silindi
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  schemas:
    CategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Kategori adı (2-100 karakter, benzersiz olmalı)
          example: "Elektronik"
        description:
          type: string
          maxLength: 500
          description: Kategori açıklaması (opsiyonel, maks 500 karakter)
          example: "Telefon, bilgisayar ve diğer elektronik ürünler"

    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Kategori benzersiz kimliği
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Kategori adı
          example: "Elektronik"
        description:
          type: string
          description: Kategori açıklaması
          example: "Telefon, bilgisayar ve diğer elektronik ürünler"
        createdAt:
          type: string
          format: date-time
          description: Oluşturulma tarihi
          example: "2026-03-11T10:00:00"
        updatedAt:
          type: string
          format: date-time
          description: Son güncelleme tarihi
          example: "2026-03-11T10:00:00"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-03-11T10:00:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Kategori bulunamadı"
        path:
          type: string
          example: "/api/v1/categories/550e8400-e29b-41d4-a716-446655440000"

  responses:
    BadRequest:
      description: Geçersiz istek verisi (validasyon hatası)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Kaynak bulunamadı
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: İsim çakışması — bu kategori adı zaten kayıtlı
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Sunucu içi hata
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"